fix cpu pegging issue + cleanup
the CPU was getting pegged because `BatchWait` was not getting set in
the Loki client. This has been now set to 1 second to reduce load.

the storage capability does not really look like it ever got
implemented, and we don't really need it as we use this as a real-time
stateless logger, so we've started to remove the storage bits.

Signed-off-by: Ryan Goodfellow <rgoodfel@isi.edu>